♻️ Split the format-neutral layer and close the format-blind doors
