+++ docs/changelog.md
@@ -18,7 +18,7 @@
 ### Added
 
 * ✨ Register `ErrorResponses()` and every rejection grelmicro raises answers the client as an [RFC 9457](https://www.rfc-editor.org/rfc/rfc9457.html) problem detail. `micro.install(app)` wires the handler on FastAPI, Starlette, and Litestar, so a rate limiter over budget, a full bulkhead, an open circuit breaker, an elapsed deadline, or an idempotency conflict returns `application/problem+json` instead of becoming a `500`. Leave the component out and grelmicro changes nothing about how your framework answers an error. ([#740](https://github.com/grelinfo/grelmicro/issues/740))
-* ✨ `grelmicro.http` ships `ProblemDetail`, a Pydantic model for the body, `problem_detail(exc)` to render a rejection, and `send_problem(send, problem)` to write one from a pure-ASGI middleware. ([#740](https://github.com/grelinfo/grelmicro/issues/740))
+* ✨ `grelmicro.http` ships `ProblemDetail` and `TMFError`, the Pydantic models for each body, `RenderedError` for what an integration writes, and `send_error(send, rendered)` to write one from a pure-ASGI middleware. ([#740](https://github.com/grelinfo/grelmicro/issues/740))
 * ✨ `CircuitBreakerError.retry_after` reports the seconds until the breaker next admits a probe, counted on the backend that holds the state so every replica agrees. It reaches a client as `Retry-After`. A breaker an operator forced open reports `0.0`, because nothing but a reset releases it. ([#740](https://github.com/grelinfo/grelmicro/issues/740))
 * ✨ `CircuitBreakerSnapshot.retry_after` carries the same value from every adapter, memory, Redis, Postgres, and SQLite alike. ([#740](https://github.com/grelinfo/grelmicro/issues/740))
 * ✨ `document_idempotency(app)` publishes a `ProblemDetail` component and points each middleware response at it, so a generated client knows the body it will get. ([#740](https://github.com/grelinfo/grelmicro/issues/740))
