+++ docs/reference/http.md
@@ -6,16 +6,15 @@
 Register `ErrorResponses()` and `micro.install(app)` wires the handler, so a
 rejection raised in a route handler answers the client as an
 `application/problem+json` body. Import `ProblemDetail` to return one of your
-own, and `send_problem` to write one from a pure-ASGI middleware.
+own, and `send_error` to write one from a pure-ASGI middleware.
 
 ::: grelmicro.http
     options:
       members:
         - ErrorResponses
+        - RenderedError
+        - send_error
         - ProblemDetail
         - TMFError
-        - RenderedError
-        - problem_detail
-        - send_problem
         - PROBLEM_MEDIA_TYPE
         - PROBLEM_TYPE_BASE
